Split credential wiring out of bootstrap egress.go
The credential-specific adapters and wiring logic (`wireCredentialResolver`,
`credentialGrantStoreLoader`, `brokerTokenResolver`, `registryMaterializer`)
move to `egress_credentials.go`. This leaves `egress.go` responsible for
policy assembly and the `EgressDeps` composition root, making it easier to
navigate as the two concerns evolve independently.
